adding 12th task and file 100-prime_factor.c

+++ more_functions_nested_loops/100-prime_factor.c
@@ -2,7 +2,7 @@
 
 int main(void)
 {
-	int num = 612852475143;
+	long long int num = 612852475143;
 	int res = -1;
 	int i;
